+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.2.0 (unreleased)
 
-* nothing here yet
+* ignore output from executor and show only stdout/stderr from actual command (#11)
 
 ## v0.1.0